Update to allow int32 for Alaska
