+++ src/delineate_hydros_and_produce_HAND.sh
@@ -311,8 +311,9 @@ if [ "$current_branch_id" = "$branch_zero_id" ] && [ "$evaluateCrosswalk" = "1" 
         -z $current_branch_id
 fi
 
-## CONVERSION TO INT16 ##
-echo -e $startDiv"Convert GW Catchments and REM to Int16 $hucNumber $current_branch_id"
-python3 $toolsDir/convert_to_int16.py \
-    -b $tempCurrentBranchDataDir
+if [ $huc2Identifier -ne 19 ]; then
+    ## CONVERSION TO INT16 ##
+    echo -e $startDiv"Convert GW Catchments and REM to Int16 $hucNumber $current_branch_id"
+    python3 $toolsDir/convert_to_int16.py \
+        -b $tempCurrentBranchDataDir
